data_work.py and hyperparameter.py created to reduce clutter. Relevant functions moved to new files. pep8 conformity performed on new files
